chore: update deps

+++ src-tauri/src/main.rs
@@ -27,6 +27,7 @@ pub struct WhisperServerDaemon(Mutex<Option<Daemon>>);
 
 fn main() {
   tauri::Builder::default()
+    .plugin(tauri_plugin_dialog::init())
     .invoke_handler(tauri::generate_handler![
       list_models,
       launch_whisper_server,
@@ -37,18 +38,14 @@ fn main() {
       get_audio_waveform_data
     ])
     .manage(WhisperServerDaemon(Default::default()))
-    .on_window_event(move |event| match event.event() {
-      WindowEvent::Destroyed => {
-        let daemon: State<'_, WhisperServerDaemon> = event.window().state();
+    .on_window_event(|window, event| {
+      if let WindowEvent::Destroyed = event {
+        let daemon: State<'_, WhisperServerDaemon> = window.state();
         let daemon = daemon.0.lock().unwrap();
-        match *daemon {
-          Some(ref daemon) => {
-            daemon.kill();
-          }
-          _ => {}
+        if let Some(ref daemon) = *daemon {
+          daemon.kill();
         }
       }
-      _ => {}
     })
     .run(tauri::generate_context!())
     .expect("error while running tauri application");
